pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "landseg"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "numpy>=1.26",
    "pillow>=10",
    "rasterio>=1.3",
    "torch>=2.2",
    "torchvision>=0.17",
    "tqdm>=4.66",
    "psutil>=5.9"
]

[project.scripts]
experiment_run = "landseg.cli.end_to_end:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["landseg*"]

[tool.setuptools.package-data]
landseg = ["configs/**/*.yaml", "configs/*.yaml"]